use Replace Constructor with Factory Function (334)
Making the Performance Calculator Polymorphic
Now that I have the logic in a class, it’s time to apply the polymorphism. The first step is
to use Replace Type Code with Subclasses (362) to introduce subclasses instead of the
type code. For this, I need to create subclasses of the performance calculator and use
the appropriate subclass in createPerformanceData. In order to get the right
subclass, I need to replace the constructor call with a function, since JavaScript
constructors can’t return subclasses. So I use Replace Constructor with Factory
Function (334).