replace temp with query(178) - play
But play is computed from the performance, so there’s no need to
pass it in as a parameter at all—I can just recalculate it within amountFor. When I’m
breaking down a long function, I like to get rid of variables like play, because
temporary variables create a lot of locally scoped names that complicate extractions.