move rend*Text args into inte*ate data str*re-cust
I now examine the other arguments used by renderPlainText. I want to move the
data that comes from them into the intermediate data structure, so that all the
calculation code moves into the statement function and renderPlainText operates
solely on data passed to it through the data parameter.
My first move is to take the customer and add it to the intermediate object (compiletestcommit).