** reorganizing the calculations by type**
Now I’ll turn my attention to the next feature change: supporting more categories of
plays, each with its own charging and volume credits calculations. At the moment, to
make changes here I have to go into the calculation functions and edit the conditions in
there. The amountFor function highlights the central role the type of play has in the
choice of calculations—but conditional logic like this tends to decay as further
modifications are made unless it’s reinforced by more structural elements of the
programming language.